requirements.txt
aiogram==2.13
torch==1.9.0
torchvision==0.10.0
tornado==6.1
tqdm==4.61.1
Pillow==8.2.0
scikit-image==0.18.1
numpy==1.20.3









